--- configs/baseline.yaml
@@ -1,39 +1,0 @@
-seed: 1822
-
-### data ###
-data_root: "data/NYU"
-remote_data_root: "/cluster_home/custom_data/NYU"
-remote_save_dir: "/job_data"
-batch_size: 1
-
-### model ###
-task: 'segmentation'
-learning_rate: 2e-5
-use_gradnorm: True
-
-### lightning ###
-lightning:
-  max_epochs: 8890
-  gpus: 0,1
-  check_val_every_n_epoch: 1
-  accelerator: "ddp"
-  num_sanity_val_steps: 0
-  progress_bar_refresh_rate: 0
-  log_every_n_steps: 5
-
-# ### traincli ###
-# traincli: 
-#   gpus: "0,1,2,3"
-#   num_machines: 1
-#   submit_root: "./"
-#   job_name: "NYUv2-seg-lr2e-5-4gpus"
-#   job_pwd: 5160
-#   project_id: "TD2021002"
-#   docker_image: "docker.hobot.cc/imagesys/hdlt:fsd_multitask-cu10-20210621-v0.3"
-#   job_list:
-#     - "pip3 install pytorch-lightning --user"
-#     - "mkdir -p /home/users/dixiao.wei/.cache/torch/hub/checkpoints"
-#     - "cp /cluster_home/custom_data/vgg16-397923af.pth /home/users/dixiao.wei/.cache/torch/hub/checkpoints/"
-#     - "python3 train.py"
-#   task_label: "pytorch"
-#   priority: 5
