+++ nevermore/__init__.py
@@ -2,3 +2,6 @@
 
 __version__ = '0.0.1'
 __author__ = ''
+
+from . import data
+from . import model
