--- nevermore/datamodule/__init__.py
@@ -1,2 +1,0 @@
-from .dataset import NUM_CLASSES, NYUv2Dateset
-from .nyuv2_datamodule import NYUv2DataModule
