--- nevermore/datamodule/dataset/__init__.py
@@ -1,1 +1,0 @@
-from .nyuv2_dataset import NUM_CLASSES, NYUv2Dateset
