+++ nevermore/model/__init__.py
@@ -1,0 +1,3 @@
+from . import layer
+from .network import SingleSegNet, MultiSegNet
+from .module import SingleSegnetNYUv2Model, MultiSegnetNYUv2Model
