+++ nevermore/model/module/__init__.py
@@ -1,0 +1,2 @@
+from .single_segnet_nyuv2 import SingleSegnetNYUv2Model
+from .multi_segnet_nyuv2 import MultiSegnetNYUv2Model
