+++ nevermore/model/network/__init__.py
@@ -1,0 +1,1 @@
+from .segnet import SingleSegNet, MultiSegNet
