+++ nevermore/model/network/segnet.py
@@ -6,6 +6,9 @@ import torch.nn as nn
 import torch.nn.functional as F
 import torchvision.models as models
 
+
+__all__ = ['SingleSegNet', 'MultiSegNet']
+
 DEBUG = False
 
 vgg16_dims = [
@@ -28,14 +31,22 @@ decoder_dims = [
 class SingleSegNet(nn.Module):
 
     def __init__(self, input_channels, output_channels):
-        super().__init__()
+        super(SingleSegNet, self).__init__()
 
         self.input_channels = input_channels
         self.output_channels = output_channels
 
         self.num_channels = input_channels
 
-        # Encoder layers
+        self.init_encoder()
+        self.init_vgg_weigts()
+        self.init_decoder()
+        self.init_head()
+
+    def init_encoder(self):
+        """
+        Create encoder layers.
+        """
 
         self.encoder_conv_00 = nn.Sequential(
             *[
@@ -175,10 +186,10 @@ class SingleSegNet(nn.Module):
             ]
         )
 
-        self.init_vgg_weigts()
-
-        # Decoder layers
-
+    def init_decoder(self):
+        """
+        Create decoder.
+        """
         self.decoder_convtr_42 = nn.Sequential(
             *[
                 nn.ConvTranspose2d(
@@ -305,6 +316,11 @@ class SingleSegNet(nn.Module):
                 nn.BatchNorm2d(64)
             ]
         )
+
+    def init_head(self):
+        """
+        Create head output layer.
+        """
         self.decoder_convtr_00 = nn.Sequential(
             *[
                 nn.ConvTranspose2d(
@@ -315,6 +331,99 @@ class SingleSegNet(nn.Module):
                 )
             ]
         )
+
+    def init_vgg_weigts(self):
+        vgg16 = models.vgg16(pretrained=True)
+        assert self.encoder_conv_00[0].weight.size(
+        ) == vgg16.features[0].weight.size()
+        self.encoder_conv_00[0].weight.data = vgg16.features[0].weight.data
+        assert self.encoder_conv_00[0].bias.size(
+        ) == vgg16.features[0].bias.size()
+        self.encoder_conv_00[0].bias.data = vgg16.features[0].bias.data
+
+        assert self.encoder_conv_01[0].weight.size(
+        ) == vgg16.features[2].weight.size()
+        self.encoder_conv_01[0].weight.data = vgg16.features[2].weight.data
+        assert self.encoder_conv_01[0].bias.size(
+        ) == vgg16.features[2].bias.size()
+        self.encoder_conv_01[0].bias.data = vgg16.features[2].bias.data
+
+        assert self.encoder_conv_10[0].weight.size(
+        ) == vgg16.features[5].weight.size()
+        self.encoder_conv_10[0].weight.data = vgg16.features[5].weight.data
+        assert self.encoder_conv_10[0].bias.size(
+        ) == vgg16.features[5].bias.size()
+        self.encoder_conv_10[0].bias.data = vgg16.features[5].bias.data
+
+        assert self.encoder_conv_11[0].weight.size(
+        ) == vgg16.features[7].weight.size()
+        self.encoder_conv_11[0].weight.data = vgg16.features[7].weight.data
+        assert self.encoder_conv_11[0].bias.size(
+        ) == vgg16.features[7].bias.size()
+        self.encoder_conv_11[0].bias.data = vgg16.features[7].bias.data
+
+        assert self.encoder_conv_20[0].weight.size(
+        ) == vgg16.features[10].weight.size()
+        self.encoder_conv_20[0].weight.data = vgg16.features[10].weight.data
+        assert self.encoder_conv_20[0].bias.size(
+        ) == vgg16.features[10].bias.size()
+        self.encoder_conv_20[0].bias.data = vgg16.features[10].bias.data
+
+        assert self.encoder_conv_21[0].weight.size(
+        ) == vgg16.features[12].weight.size()
+        self.encoder_conv_21[0].weight.data = vgg16.features[12].weight.data
+        assert self.encoder_conv_21[0].bias.size(
+        ) == vgg16.features[12].bias.size()
+        self.encoder_conv_21[0].bias.data = vgg16.features[12].bias.data
+
+        assert self.encoder_conv_22[0].weight.size(
+        ) == vgg16.features[14].weight.size()
+        self.encoder_conv_22[0].weight.data = vgg16.features[14].weight.data
+        assert self.encoder_conv_22[0].bias.size(
+        ) == vgg16.features[14].bias.size()
+        self.encoder_conv_22[0].bias.data = vgg16.features[14].bias.data
+
+        assert self.encoder_conv_30[0].weight.size(
+        ) == vgg16.features[17].weight.size()
+        self.encoder_conv_30[0].weight.data = vgg16.features[17].weight.data
+        assert self.encoder_conv_30[0].bias.size(
+        ) == vgg16.features[17].bias.size()
+        self.encoder_conv_30[0].bias.data = vgg16.features[17].bias.data
+
+        assert self.encoder_conv_31[0].weight.size(
+        ) == vgg16.features[19].weight.size()
+        self.encoder_conv_31[0].weight.data = vgg16.features[19].weight.data
+        assert self.encoder_conv_31[0].bias.size(
+        ) == vgg16.features[19].bias.size()
+        self.encoder_conv_31[0].bias.data = vgg16.features[19].bias.data
+
+        assert self.encoder_conv_32[0].weight.size(
+        ) == vgg16.features[21].weight.size()
+        self.encoder_conv_32[0].weight.data = vgg16.features[21].weight.data
+        assert self.encoder_conv_32[0].bias.size(
+        ) == vgg16.features[21].bias.size()
+        self.encoder_conv_32[0].bias.data = vgg16.features[21].bias.data
+
+        assert self.encoder_conv_40[0].weight.size(
+        ) == vgg16.features[24].weight.size()
+        self.encoder_conv_40[0].weight.data = vgg16.features[24].weight.data
+        assert self.encoder_conv_40[0].bias.size(
+        ) == vgg16.features[24].bias.size()
+        self.encoder_conv_40[0].bias.data = vgg16.features[24].bias.data
+
+        assert self.encoder_conv_41[0].weight.size(
+        ) == vgg16.features[26].weight.size()
+        self.encoder_conv_41[0].weight.data = vgg16.features[26].weight.data
+        assert self.encoder_conv_41[0].bias.size(
+        ) == vgg16.features[26].bias.size()
+        self.encoder_conv_41[0].bias.data = vgg16.features[26].bias.data
+
+        assert self.encoder_conv_42[0].weight.size(
+        ) == vgg16.features[28].weight.size()
+        self.encoder_conv_42[0].weight.data = vgg16.features[28].weight.data
+        assert self.encoder_conv_42[0].bias.size(
+        ) == vgg16.features[28].bias.size()
+        self.encoder_conv_42[0].bias.data = vgg16.features[28].bias.data
 
     def forward(self, input_img):
         """
@@ -431,95 +540,168 @@ class SingleSegNet(nn.Module):
 
         return x_00d, x_softmax
 
-    def init_vgg_weigts(self):
-        vgg16 = models.vgg16(pretrained=True)
-        assert self.encoder_conv_00[0].weight.size(
-        ) == vgg16.features[0].weight.size()
-        self.encoder_conv_00[0].weight.data = vgg16.features[0].weight.data
-        assert self.encoder_conv_00[0].bias.size(
-        ) == vgg16.features[0].bias.size()
-        self.encoder_conv_00[0].bias.data = vgg16.features[0].bias.data
 
-        assert self.encoder_conv_01[0].weight.size(
-        ) == vgg16.features[2].weight.size()
-        self.encoder_conv_01[0].weight.data = vgg16.features[2].weight.data
-        assert self.encoder_conv_01[0].bias.size(
-        ) == vgg16.features[2].bias.size()
-        self.encoder_conv_01[0].bias.data = vgg16.features[2].bias.data
+class MultiSegNet(SingleSegNet):
 
-        assert self.encoder_conv_10[0].weight.size(
-        ) == vgg16.features[5].weight.size()
-        self.encoder_conv_10[0].weight.data = vgg16.features[5].weight.data
-        assert self.encoder_conv_10[0].bias.size(
-        ) == vgg16.features[5].bias.size()
-        self.encoder_conv_10[0].bias.data = vgg16.features[5].bias.data
+    def __init__(
+        self,
+        input_channels,
+        seg_output_channels,
+        dep_output_channels,
+        nor_output_channels,
+    ):
 
-        assert self.encoder_conv_11[0].weight.size(
-        ) == vgg16.features[7].weight.size()
-        self.encoder_conv_11[0].weight.data = vgg16.features[7].weight.data
-        assert self.encoder_conv_11[0].bias.size(
-        ) == vgg16.features[7].bias.size()
-        self.encoder_conv_11[0].bias.data = vgg16.features[7].bias.data
+        self.seg_output_channels = seg_output_channels
+        self.dep_output_channels = dep_output_channels
+        self.nor_output_channels = nor_output_channels
+        super(MultiSegNet, self).__init__(input_channels, None)
 
-        assert self.encoder_conv_20[0].weight.size(
-        ) == vgg16.features[10].weight.size()
-        self.encoder_conv_20[0].weight.data = vgg16.features[10].weight.data
-        assert self.encoder_conv_20[0].bias.size(
-        ) == vgg16.features[10].bias.size()
-        self.encoder_conv_20[0].bias.data = vgg16.features[10].bias.data
+    def init_head(self):
+        self.decoder_convtr_00_seg = nn.Sequential(
+            *[
+                nn.ConvTranspose2d(
+                    in_channels=64,
+                    out_channels=self.seg_output_channels,
+                    kernel_size=3,
+                    padding=1
+                )
+            ]
+        )
+        self.decoder_convtr_00_dep = nn.Sequential(
+            *[
+                nn.ConvTranspose2d(
+                    in_channels=64,
+                    out_channels=self.dep_output_channels,
+                    kernel_size=3,
+                    padding=1
+                )
+            ]
+        )
+        self.decoder_convtr_00_nor = nn.Sequential(
+            *[
+                nn.ConvTranspose2d(
+                    in_channels=64,
+                    out_channels=self.nor_output_channels,
+                    kernel_size=3,
+                    padding=1
+                )
+            ]
+        )
 
-        assert self.encoder_conv_21[0].weight.size(
-        ) == vgg16.features[12].weight.size()
-        self.encoder_conv_21[0].weight.data = vgg16.features[12].weight.data
-        assert self.encoder_conv_21[0].bias.size(
-        ) == vgg16.features[12].bias.size()
-        self.encoder_conv_21[0].bias.data = vgg16.features[12].bias.data
+    def forward(self, input_img):
+        """
+        Forward pass `input_img` through the network
 
-        assert self.encoder_conv_22[0].weight.size(
-        ) == vgg16.features[14].weight.size()
-        self.encoder_conv_22[0].weight.data = vgg16.features[14].weight.data
-        assert self.encoder_conv_22[0].bias.size(
-        ) == vgg16.features[14].bias.size()
-        self.encoder_conv_22[0].bias.data = vgg16.features[14].bias.data
+        # TODO: Refactor with SingleSegNet <31-08-21, Duino> #
+        """
+        # Encoder
 
-        assert self.encoder_conv_30[0].weight.size(
-        ) == vgg16.features[17].weight.size()
-        self.encoder_conv_30[0].weight.data = vgg16.features[17].weight.data
-        assert self.encoder_conv_30[0].bias.size(
-        ) == vgg16.features[17].bias.size()
-        self.encoder_conv_30[0].bias.data = vgg16.features[17].bias.data
+        # Encoder Stage - 1
+        dim_0 = input_img.size()
+        x_00 = F.relu(self.encoder_conv_00(input_img))
+        x_01 = F.relu(self.encoder_conv_01(x_00))
+        x_0, indices_0 = F.max_pool2d(
+            x_01, kernel_size=2, stride=2, return_indices=True
+        )
 
-        assert self.encoder_conv_31[0].weight.size(
-        ) == vgg16.features[19].weight.size()
-        self.encoder_conv_31[0].weight.data = vgg16.features[19].weight.data
-        assert self.encoder_conv_31[0].bias.size(
-        ) == vgg16.features[19].bias.size()
-        self.encoder_conv_31[0].bias.data = vgg16.features[19].bias.data
+        # Encoder Stage - 2
+        dim_1 = x_0.size()
+        x_10 = F.relu(self.encoder_conv_10(x_0))
+        x_11 = F.relu(self.encoder_conv_11(x_10))
+        x_1, indices_1 = F.max_pool2d(
+            x_11, kernel_size=2, stride=2, return_indices=True
+        )
 
-        assert self.encoder_conv_32[0].weight.size(
-        ) == vgg16.features[21].weight.size()
-        self.encoder_conv_32[0].weight.data = vgg16.features[21].weight.data
-        assert self.encoder_conv_32[0].bias.size(
-        ) == vgg16.features[21].bias.size()
-        self.encoder_conv_32[0].bias.data = vgg16.features[21].bias.data
+        # Encoder Stage - 3
+        dim_2 = x_1.size()
+        x_20 = F.relu(self.encoder_conv_20(x_1))
+        x_21 = F.relu(self.encoder_conv_21(x_20))
+        x_22 = F.relu(self.encoder_conv_22(x_21))
+        x_2, indices_2 = F.max_pool2d(
+            x_22, kernel_size=2, stride=2, return_indices=True
+        )
 
-        assert self.encoder_conv_40[0].weight.size(
-        ) == vgg16.features[24].weight.size()
-        self.encoder_conv_40[0].weight.data = vgg16.features[24].weight.data
-        assert self.encoder_conv_40[0].bias.size(
-        ) == vgg16.features[24].bias.size()
-        self.encoder_conv_40[0].bias.data = vgg16.features[24].bias.data
+        # Encoder Stage - 4
+        dim_3 = x_2.size()
+        x_30 = F.relu(self.encoder_conv_30(x_2))
+        x_31 = F.relu(self.encoder_conv_31(x_30))
+        x_32 = F.relu(self.encoder_conv_32(x_31))
+        x_3, indices_3 = F.max_pool2d(
+            x_32, kernel_size=2, stride=2, return_indices=True
+        )
 
-        assert self.encoder_conv_41[0].weight.size(
-        ) == vgg16.features[26].weight.size()
-        self.encoder_conv_41[0].weight.data = vgg16.features[26].weight.data
-        assert self.encoder_conv_41[0].bias.size(
-        ) == vgg16.features[26].bias.size()
-        self.encoder_conv_41[0].bias.data = vgg16.features[26].bias.data
+        # Encoder Stage - 5
+        dim_4 = x_3.size()
+        x_40 = F.relu(self.encoder_conv_40(x_3))
+        x_41 = F.relu(self.encoder_conv_41(x_40))
+        x_42 = F.relu(self.encoder_conv_42(x_41))
+        x_4, indices_4 = F.max_pool2d(
+            x_42, kernel_size=2, stride=2, return_indices=True
+        )
 
-        assert self.encoder_conv_42[0].weight.size(
-        ) == vgg16.features[28].weight.size()
-        self.encoder_conv_42[0].weight.data = vgg16.features[28].weight.data
-        assert self.encoder_conv_42[0].bias.size(
-        ) == vgg16.features[28].bias.size()
-        self.encoder_conv_42[0].bias.data = vgg16.features[28].bias.data
+        # Decoder
+
+        dim_d = x_4.size()
+
+        # Decoder Stage - 5
+        x_4d = F.max_unpool2d(
+            x_4, indices_4, kernel_size=2, stride=2, output_size=dim_4
+        )
+        x_42d = F.relu(self.decoder_convtr_42(x_4d))
+        x_41d = F.relu(self.decoder_convtr_41(x_42d))
+        x_40d = F.relu(self.decoder_convtr_40(x_41d))
+        dim_4d = x_40d.size()
+
+        # Decoder Stage - 4
+        x_3d = F.max_unpool2d(
+            x_40d, indices_3, kernel_size=2, stride=2, output_size=dim_3
+        )
+        x_32d = F.relu(self.decoder_convtr_32(x_3d))
+        x_31d = F.relu(self.decoder_convtr_31(x_32d))
+        x_30d = F.relu(self.decoder_convtr_30(x_31d))
+        dim_3d = x_30d.size()
+
+        # Decoder Stage - 3
+        x_2d = F.max_unpool2d(
+            x_30d, indices_2, kernel_size=2, stride=2, output_size=dim_2
+        )
+        x_22d = F.relu(self.decoder_convtr_22(x_2d))
+        x_21d = F.relu(self.decoder_convtr_21(x_22d))
+        x_20d = F.relu(self.decoder_convtr_20(x_21d))
+        dim_2d = x_20d.size()
+
+        # Decoder Stage - 2
+        x_1d = F.max_unpool2d(
+            x_20d, indices_1, kernel_size=2, stride=2, output_size=dim_1
+        )
+        x_11d = F.relu(self.decoder_convtr_11(x_1d))
+        x_10d = F.relu(self.decoder_convtr_10(x_11d))
+        dim_1d = x_10d.size()
+
+        # Decoder Stage - 1
+        x_0d = F.max_unpool2d(
+            x_10d, indices_0, kernel_size=2, stride=2, output_size=dim_0
+        )
+        x_01d = F.relu(self.decoder_convtr_01(x_0d))
+        x_00d_seg = self.decoder_convtr_00_seg(x_01d)
+        x_00d_dep = self.decoder_convtr_00_dep(x_01d)
+        x_00d_nor = self.decoder_convtr_00_nor(x_01d)
+        dim_0d = x_00d_seg.size()
+
+        x_softmax_seg = F.softmax(x_00d_seg, dim=1)
+
+        if DEBUG:
+            print("dim_0: {}".format(dim_0))
+            print("dim_1: {}".format(dim_1))
+            print("dim_2: {}".format(dim_2))
+            print("dim_3: {}".format(dim_3))
+            print("dim_4: {}".format(dim_4))
+
+            print("dim_d: {}".format(dim_d))
+            print("dim_4d: {}".format(dim_4d))
+            print("dim_3d: {}".format(dim_3d))
+            print("dim_2d: {}".format(dim_2d))
+            print("dim_1d: {}".format(dim_1d))
+            print("dim_0d: {}".format(dim_0d))
+
+        return x_00d_seg, x_00d_dep, x_00d_nor, x_softmax_seg
