--- nevermore/module/__init__.py
@@ -1,3 +1,0 @@
-from .layers import *
-from .model import *
-from .single_segnet_nyuv2 import SingleSegnetNyuv2Model
