--- nevermore/module/model/__init__.py
@@ -1,2 +1,0 @@
-from .multi_segnet import MultiSegNet
-from .single_segnet import SingleSegNet
